Add MyCanvas class
Played around with the difference between hiding the navigation bar,
and setting SYSTEM_UI_FLAG_FULLSCREEN flag for the systemUivisibility of the whole view.
Was neat to see that the flag gets rid of the background where the inset is, that would
still show up if we had only taken out the navigation bar. :100: