Don't store path in bitmap, store in Path instead
Previously we were caching the map in a bitmap and storing an extra canvas
as well. We switched over to storing the path in a single 'curPath' variable
and the store the entire drawing, which is an accumulation of the 'curPath' every
time we move our finger back off the screen.